Move dtypes into shared
PSG will have shared data types, early refactor
