+++ src/MAKE_ATR
@@ -17,6 +17,7 @@ TESTS_FOLDER = tests
 EVENTS_FOLDER = events
 INPUT_FOLDER = input
 CLOCK_FOLDER = clock
+SHARED_FOLDER = shared
 TESTS_MODEL_FOLDER = $(TESTS_FOLDER)\$(MODEL_FOLDER)
 TESTS_RASTER_FOLDER = $(TESTS_FOLDER)\$(RASTER_FOLDER)
 TESTS_RENDER_FOLDER = $(TESTS_FOLDER)\$(RENDER_FOLDER)
@@ -74,7 +75,7 @@ ATARI_CC     = cc68x
 SHARED_FLAGS = -g
 
 # Program Configuration
-TOS_FLAGS      = $(SHARED_FLAGS) -c -I. -I$(MODEL_FOLDER) -I$(RASTER_FOLDER) -I$(RENDER_FOLDER) -I$(EVENTS_FOLDER) -I$(INPUT_FOLDER) -I$(CLOCK_FOLDER)
+TOS_FLAGS      = $(SHARED_FLAGS) -c -I. -I$(MODEL_FOLDER) -I$(SHARED_FOLDER) -I$(RASTER_FOLDER) -I$(RENDER_FOLDER) -I$(EVENTS_FOLDER) -I$(INPUT_FOLDER) -I$(CLOCK_FOLDER)
 MODEL_OBJECTS  = $(BLOCK).o $(SPIKE).o $(LAVA).o $(GEO).o $(LEVEL).o $(WORLD).o $(MODEL).o $(CAMERA).o $(MAIN).o
 EVENTS_OBJECTS = $(ASYNC).o $(COND).o $(SYNC).o
 CLOCK_OBJECTS = $(CLOCK).o
@@ -85,16 +86,16 @@ TOS_OBJECTS    = $(MODEL_OBJECTS) $(RASTER_OBJECTS) $(RENDER_OBJECTS) $(EVENTS_O
 
 TOS_NAME       = dash.tos
 # TODO: Feels redundant
-MAIN_FLAGS     = $(SHARED_FLAGS) -I. -I$(MODEL_FOLDER) -I$(RASTER_FOLDER) -I$(RENDER_FOLDER) -I$(EVENTS_FOLDER) -I$(INPUT_FOLDER)
+MAIN_FLAGS     = $(SHARED_FLAGS) -I. -I$(MODEL_FOLDER) -I$(SHARED_FOLDER) -I$(RASTER_FOLDER) -I$(RENDER_FOLDER) -I$(EVENTS_FOLDER) -I$(INPUT_FOLDER)
 
 # Test Configuration
 # TODO: It's a bit silly, but pretty sure we're including all our .c files for each test.
-TEST_FLAGS = $(SHARED_FLAGS) -I. -I$(MODEL_FOLDER) -I$(RENDER_FOLDER) -I$(UNITY_FOLDER) -DUNITY_EXCLUDE_STDINT_H -DUNITY_INT_WIDTH=16 -DUNITY_LONG_WIDTH=32 -DUNITY_POINTER_WIDTH=32 -DTESTING
+TEST_FLAGS = $(SHARED_FLAGS) -I. -I$(MODEL_FOLDER) -I$(SHARED_FOLDER) -I$(RENDER_FOLDER) -I$(UNITY_FOLDER) -DUNITY_EXCLUDE_STDINT_H -DUNITY_INT_WIDTH=16 -DUNITY_LONG_WIDTH=32 -DUNITY_POINTER_WIDTH=32 -DTESTING
 TEST_SRC   = $(MODEL_FOLDER)\$(BLOCK).c $(MODEL_FOLDER)\$(SPIKE).c $(MODEL_FOLDER)\$(LAVA).c $(MODEL_FOLDER)\$(GEO).c $(MODEL_FOLDER)\$(LEVEL).c $(MODEL_FOLDER)\$(WORLD).c $(MODEL_FOLDER)\$(MODEL).c $(MODEL_FOLDER)\$(CAMERA).c
 UNITY      = $(UNITY_FOLDER)\unity.c
 
 # Events Test Configuration
-EVENTS_TEST_FLAGS = $(SHARED_FLAGS) -I. -I$(MODEL_FOLDER) -I$(RENDER_FOLDER) -I$(EVENTS_FOLDER) -I$(UNITY_FOLDER) -DUNITY_EXCLUDE_STDINT_H -DUNITY_INT_WIDTH=16 -DUNITY_LONG_WIDTH=32 -DUNITY_POINTER_WIDTH=32 -DTESTING
+EVENTS_TEST_FLAGS = $(SHARED_FLAGS) -I. -I$(MODEL_FOLDER) -I$(SHARED_FOLDER) -I$(RENDER_FOLDER) -I$(EVENTS_FOLDER) -I$(UNITY_FOLDER) -DUNITY_EXCLUDE_STDINT_H -DUNITY_INT_WIDTH=16 -DUNITY_LONG_WIDTH=32 -DUNITY_POINTER_WIDTH=32 -DTESTING
 EVENTS_TEST_SRC = $(TEST_SRC) $(EVENTS_FOLDER)\$(ASYNC).c $(EVENTS_FOLDER)\$(COND).c $(EVENTS_FOLDER)\$(SYNC).c
 
 # Raster Test Configuration
@@ -104,7 +105,7 @@ RASTER_TEST_FLAGS = $(SHARED_FLAGS) -I. -I$(RASTER_FOLDER) -I$(UNITY_FOLDER) -DU
 CLOCK_TEST_FLAGS = $(SHARED_FLAGS) -I. -I$(CLOCK_FOLDER) -I$(UNITY_FOLDER) -DUNITY_EXCLUDE_STDINT_H -DUNITY_INT_WIDTH=16 -DUNITY_LONG_WIDTH=32 -DUNITY_POINTER_WIDTH=32 -DTESTING
 
 # Render Test Configuration
-RENDER_TEST_FLAGS = $(SHARED_FLAGS) -I. -I$(MODEL_FOLDER) -I$(RENDER_FOLDER) -I$(UNITY_FOLDER) -DUNITY_EXCLUDE_STDINT_H -DUNITY_INT_WIDTH=16 -DUNITY_LONG_WIDTH=32 -DUNITY_POINTER_WIDTH=32 -DTESTING
+RENDER_TEST_FLAGS = $(SHARED_FLAGS) -I. -I$(MODEL_FOLDER) -I$(SHARED_FOLDER) -I$(RENDER_FOLDER) -I$(UNITY_FOLDER) -DUNITY_EXCLUDE_STDINT_H -DUNITY_INT_WIDTH=16 -DUNITY_LONG_WIDTH=32 -DUNITY_POINTER_WIDTH=32 -DTESTING
 RENDER_TEST_SRC = $(TEST_SRC) $(RENDER_FOLDER)\$(RENDER).c
 
 # File Names, Tests - Events
@@ -153,46 +154,46 @@ TCLOCK = t${CLOCK}
 $(TOS_NAME): $(TOS_OBJECTS)
 	$(ATARI_CC) $(MAIN_FLAGS) $(TOS_OBJECTS) -o $(TOS_NAME)
 
-$(CAMERA).o: $(MODEL_FOLDER)\$(CAMERA).c $(MODEL_FOLDER)\$(CAMERA).h $(MODEL_FOLDER)\dtypes.h
+$(CAMERA).o: $(MODEL_FOLDER)\$(CAMERA).c $(MODEL_FOLDER)\$(CAMERA).h $(SHARED_FOLDER)\dtypes.h
 	$(ATARI_CC) $(TOS_FLAGS) $(MODEL_FOLDER)\$(CAMERA).c -o $(CAMERA).o
 
-$(BLOCK).o: $(MODEL_FOLDER)\$(BLOCK).c $(MODEL_FOLDER)\$(BLOCK).h $(MODEL_FOLDER)\dtypes.h
+$(BLOCK).o: $(MODEL_FOLDER)\$(BLOCK).c $(MODEL_FOLDER)\$(BLOCK).h $(SHARED_FOLDER)\dtypes.h
 	$(ATARI_CC) $(TOS_FLAGS) $(MODEL_FOLDER)\$(BLOCK).c -o $(BLOCK).o
 
-$(LAVA).o: $(MODEL_FOLDER)\$(LAVA).c $(MODEL_FOLDER)\$(LAVA).h $(MODEL_FOLDER)\dtypes.h
+$(LAVA).o: $(MODEL_FOLDER)\$(LAVA).c $(MODEL_FOLDER)\$(LAVA).h $(SHARED_FOLDER)\dtypes.h
 	$(ATARI_CC) $(TOS_FLAGS) $(MODEL_FOLDER)\$(LAVA).c -o $(LAVA).o
 
-$(SPIKE).o: $(MODEL_FOLDER)\$(SPIKE).c $(MODEL_FOLDER)\$(SPIKE).h $(MODEL_FOLDER)\dtypes.h
+$(SPIKE).o: $(MODEL_FOLDER)\$(SPIKE).c $(MODEL_FOLDER)\$(SPIKE).h $(SHARED_FOLDER)\dtypes.h
 	$(ATARI_CC) $(TOS_FLAGS) $(MODEL_FOLDER)\$(SPIKE).c -o $(SPIKE).o
 
-$(GEO).o: $(MODEL_FOLDER)\$(GEO).c $(MODEL_FOLDER)\$(GEO).h $(MODEL_FOLDER)\dtypes.h
+$(GEO).o: $(MODEL_FOLDER)\$(GEO).c $(MODEL_FOLDER)\$(GEO).h $(SHARED_FOLDER)\dtypes.h
 	$(ATARI_CC) $(TOS_FLAGS) $(MODEL_FOLDER)\$(GEO).c -o $(GEO).o
 
-$(LEVEL).o: $(MODEL_FOLDER)\$(LEVEL).c $(MODEL_FOLDER)\$(LEVEL).h $(MODEL_FOLDER)\dtypes.h
+$(LEVEL).o: $(MODEL_FOLDER)\$(LEVEL).c $(MODEL_FOLDER)\$(LEVEL).h $(SHARED_FOLDER)\dtypes.h
 	$(ATARI_CC) $(TOS_FLAGS) $(MODEL_FOLDER)\$(LEVEL).c -o $(LEVEL).o
 
-$(WORLD).o: $(MODEL_FOLDER)\$(WORLD).c $(MODEL_FOLDER)\$(WORLD).h $(MODEL_FOLDER)\dtypes.h
+$(WORLD).o: $(MODEL_FOLDER)\$(WORLD).c $(MODEL_FOLDER)\$(WORLD).h $(SHARED_FOLDER)\dtypes.h
 	$(ATARI_CC) $(TOS_FLAGS) $(MODEL_FOLDER)\$(WORLD).c -o $(WORLD).o
 
-$(MODEL).o: $(MODEL_FOLDER)\$(MODEL).c $(MODEL_FOLDER)\$(MODEL).h $(MODEL_FOLDER)\dtypes.h
+$(MODEL).o: $(MODEL_FOLDER)\$(MODEL).c $(MODEL_FOLDER)\$(MODEL).h $(SHARED_FOLDER)\dtypes.h
 	$(ATARI_CC) $(TOS_FLAGS) $(MODEL_FOLDER)\$(MODEL).c -o $(MODEL).o
 
-$(MAIN).o: $(MAIN).c $(MODEL_FOLDER)\$(MODEL).h $(MODEL_FOLDER)\dtypes.h $(RASTER_FOLDER)\$(RASTER).h $(CLOCK_FOLDER)\$(CLOCK).h $(INPUT_FOLDER)\$(INPUT).h
+$(MAIN).o: $(MAIN).c $(MODEL_FOLDER)\$(MODEL).h $(SHARED_FOLDER)\dtypes.h $(RASTER_FOLDER)\$(RASTER).h $(CLOCK_FOLDER)\$(CLOCK).h $(INPUT_FOLDER)\$(INPUT).h
 	$(ATARI_CC) $(TOS_FLAGS) $(MAIN).c -o $(MAIN).o
 
-$(ASYNC).o: $(EVENTS_FOLDER)\$(ASYNC).c $(EVENTS_FOLDER)\$(ASYNC).h $(MODEL_FOLDER)\dtypes.h
+$(ASYNC).o: $(EVENTS_FOLDER)\$(ASYNC).c $(EVENTS_FOLDER)\$(ASYNC).h $(SHARED_FOLDER)\dtypes.h
 	$(ATARI_CC) $(TOS_FLAGS) $(EVENTS_FOLDER)\$(ASYNC).c -o $(ASYNC).o
 
-$(COND).o: $(EVENTS_FOLDER)\$(COND).c $(EVENTS_FOLDER)\$(COND).h $(MODEL_FOLDER)\dtypes.h
+$(COND).o: $(EVENTS_FOLDER)\$(COND).c $(EVENTS_FOLDER)\$(COND).h $(SHARED_FOLDER)\dtypes.h
 	$(ATARI_CC) $(TOS_FLAGS) $(EVENTS_FOLDER)\$(COND).c -o $(COND).o
 
-$(SYNC).o: $(EVENTS_FOLDER)\$(SYNC).c $(EVENTS_FOLDER)\$(SYNC).h $(MODEL_FOLDER)\dtypes.h
+$(SYNC).o: $(EVENTS_FOLDER)\$(SYNC).c $(EVENTS_FOLDER)\$(SYNC).h $(SHARED_FOLDER)\dtypes.h
 	$(ATARI_CC) $(TOS_FLAGS) $(EVENTS_FOLDER)\$(SYNC).c -o $(SYNC).o
 
-$(INPUT).o: $(INPUT_FOLDER)\$(INPUT).c $(INPUT_FOLDER)\$(INPUT).h $(MODEL_FOLDER)\dtypes.h
+$(INPUT).o: $(INPUT_FOLDER)\$(INPUT).c $(INPUT_FOLDER)\$(INPUT).h $(SHARED_FOLDER)\dtypes.h
 	$(ATARI_CC) $(TOS_FLAGS) $(INPUT_FOLDER)\$(INPUT).c -o $(INPUT).o
 
-$(RENDER).o: $(RENDER_FOLDER)\$(RENDER).c $(RENDER_FOLDER)\$(RENDER).h $(MODEL_FOLDER)\dtypes.h $(RASTER_FOLDER)\$(RASTER).h
+$(RENDER).o: $(RENDER_FOLDER)\$(RENDER).c $(RENDER_FOLDER)\$(RENDER).h $(SHARED_FOLDER)\dtypes.h $(RASTER_FOLDER)\$(RASTER).h
 	$(ATARI_CC) $(TOS_FLAGS) $(RENDER_FOLDER)\$(RENDER).c -o $(RENDER).o
 
 $(CLOCK).o: $(CLOCK_FOLDER)\$(CLOCK).c $(CLOCK_FOLDER)\$(CLOCK).h
